Add VimixTestCurrentLine
This will work like VimixTestCurrentFile, but append the current line number so
a test can be conveniently run in isolation.

It will produce an error when we are not currently in a test file, because in
this case we cannot really determine which line from the test file should run.